// src/app/Models/UserModel.ts
export interface UserModel {
  idStoreUser: number;
  idShop: number;
  uuid: string;
  code: string;
  shopName: string;
  idOffice: any;
  officeName: string;
  document: string;
  name: string;
  lastName: string;
  email: string;
  numberPhone: string;
  username: string;
  password: string;
  status: string;
  idRol: number;
  rolName: string;
  options: string;
  userActions: string;
  lastLoginDate: string;
  dateCreated: string;
  onLine: string;
  dateUpdated: string;
}